Remove .padEnd (#8)
We should figure out what interpreters we're targeting / compatibility
issues. We can add this back in, but for now I'm just gonna do the easy
thing and say "This doesn't work on my machine" and rip it out. For now
we don't need the padding anyway, since we have this fixed date format.